Refactor AuthResultInterface implementations
